Adds a way to se the default language and separates files
